properly handle integer overflow in sendSCPStateToPeer

+++ src/historywork/FetchRecentQsetsWork.cpp
@@ -72,8 +72,7 @@ FetchRecentQsetsWork::onSuccess()
     }
 
     // Phase 3: extract the qsets.
-    // use uint64_t for i to prevent overflows
-    for (uint64_t i = firstSeq; i <= lastSeq; i += step)
+    for (uint32_t i = firstSeq; i <= lastSeq; i += step)
     {
         CLOG(INFO, "History") << "Scanning for QSets in checkpoint: " << i;
         XDRInputFileStream in;
